+++ ExpandWorldEvents/SpawnLoader.cs
@@ -8,8 +8,9 @@ using UnityEngine;
 namespace ExpandWorld.Spawn;
 public class Loader
 {
-  public static Dictionary<SpawnSystem.SpawnData, ZDOData?> Data = new();
+  public static Dictionary<SpawnSystem.SpawnData, ZDOData?> Data = [];
   public static Dictionary<SpawnSystem.SpawnData, List<BlueprintObject>> Objects = [];
+  private static readonly int HashFaction = "faction".GetStableHashCode();
 
   public static SpawnSystem.SpawnData FromData(Data data)
   {
@@ -45,13 +46,22 @@ public class Loader
       m_maxLevel = data.maxLevel,
       m_minLevel = data.minLevel,
       m_levelUpMinCenterDistance = data.levelUpMinCenterDistance,
-      m_overrideLevelupChance = data.overrideLevelupChance
+      m_overrideLevelupChance = data.overrideLevelupChance,
     };
     if (spawn.m_minAltitude == -10000f)
       spawn.m_minAltitude = spawn.m_maxAltitude > 0f ? 0f : -1000f;
     if (data.data != "")
     {
       Data[spawn] = ZDOData.Create(data.data);
+    }
+    if (data.faction != "")
+    {
+      var factionData = new ZDOData();
+      factionData.Strings[HashFaction] = data.faction;
+      if (Data.ContainsKey(spawn))
+        Data[spawn] = ZDOData.Merge(Data[spawn], factionData);
+      else
+        Data.Add(spawn, factionData);
     }
     if (data.objects != null)
     {